Create List2
Read 10 money amounts from the user and store these in a list. For each amount, make a change with coins in [0.01, 0.02, 0.05, 0.10, 0.20, 0.50, 1, 2] euros. For each amount, calculate the minimum number of coins needed for the change and keep them in a tuple for each original amount.

For example, if the amount is 20.77 then the best solution will be 10x2 + 0.50x1 + 0.20x1 + 0.05x1 + 0.02x1